Split up script into simpler files
